Make the parser strict, report line:col errors, add --trace
- Follow the RFC 8259 grammar exactly: required commas and colons,
  no trailing commas, number grammar with exponents and no leading
  zeros, all escapes incl. \uXXXX surrogate pairs, no raw control
  characters, no trailing garbage, nesting limit of 512.
- Throw JsonParseError with byte offset and 1-based line/column
  instead of returning nullptr or reading past the buffer.
- Optional JsonTrace records enter/exit of each grammar rule, tokens,
  node creation and the error, for the web visualizer.
- JsonNumber stores double (was float); JsonObject keeps insertion order.
- main.cpp: CLI that pretty-prints a file or stdin, --trace, --demo
  (the original examples); printer escapes strings.

+++ JsonArray.h
@@ -14,7 +14,7 @@ public:
 		members.push_back(std::move(value));
 	};
 
-	int size() const { return members.size(); };
+	int size() const { return static_cast<int>(members.size()); };
 
 	JsonValue& operator[] (int index) {
 		return *members[index];
