+++ JsonNumber.h
@@ -5,11 +5,10 @@
 class JsonNumber: public JsonValue
 {
 public:
-	JsonNumber(float v): value(v) {
+	JsonNumber(double v): value(v) {
 		this->setType(JsonType::NUMBER);
 	};
-	float getValue() const { return value; };
+	double getValue() const { return value; };
 private:
-	float value;
+	double value;
 };
-
